+++ src/components/ConsortiaControlledVocabulary/index.js
@@ -1,0 +1,1 @@
+export { ConsortiaControlledVocabulary } from './ConsortiaControlledVocabulary';