+++ src/components/ConsortiaControlledVocabulary/modals/index.js
@@ -1,0 +1,2 @@
+export { ConfirmDeleteEntryModal } from './ConfirmDeleteEntryModal';
+export { ItemInUseModal } from './ItemInUseModal';