+++ src/routes/ConsortiumManager/settings/users/general/PatronGroups/index.js
@@ -1,0 +1,1 @@
+export { PatronGroups } from './PatronGroups';